+++ src/main/java/io/github/torand/fastersql/function/aggregate/Sum.java
@@ -17,8 +17,10 @@ package io.github.torand.fastersql.function.aggregate;
 
 import io.github.torand.fastersql.Context;
 import io.github.torand.fastersql.Field;
+import io.github.torand.fastersql.expression.Expression;
 import io.github.torand.fastersql.projection.Projection;
 
+import java.util.Random;
 import java.util.stream.Stream;
 
 import static io.github.torand.fastersql.util.contract.Requires.requireNonBlank;
@@ -26,19 +28,19 @@ import static io.github.torand.fastersql.util.lang.StringHelper.nonBlank;
 import static java.util.Objects.requireNonNull;
 
 public class Sum implements AggregateFunction {
-    private final Field field;
+    private final Expression expression;
     private final String alias;
 
-    Sum(Field field, String alias) {
-        this.field = requireNonNull(field, "No field specified");
-        this.alias = nonBlank(alias) ? alias : defaultAlias(field);
+    Sum(Expression expression, String alias) {
+        this.expression = requireNonNull(expression, "No expression specified");
+        this.alias = nonBlank(alias) ? alias : defaultAlias();
     }
 
     // Sql
 
     @Override
     public String sql(Context context) {
-        return "sum(" + field.sql(context) + ")";
+        return "sum(" + expression.sql(context) + ")";
     }
 
     @Override
@@ -51,7 +53,7 @@ public class Sum implements AggregateFunction {
     @Override
     public Projection as(String alias) {
         requireNonBlank(alias, "No alias specified");
-        return new Sum(field, alias);
+        return new Sum(expression, alias);
     }
 
     @Override
@@ -63,10 +65,10 @@ public class Sum implements AggregateFunction {
 
     @Override
     public Stream<Field> fieldRefs() {
-        return Stream.of(field);
+        return expression.fieldRefs();
     }
 
-    private String defaultAlias(Field field) {
-        return "SUM_" + field.alias();
+    private String defaultAlias() {
+        return "SUM_" + new Random().nextInt(999) + 1;
     }
 }