+++ src/main/java/io/github/torand/fastersql/function/singlerow/Lower.java
@@ -45,7 +45,7 @@ public class Lower implements SingleRowFunction {
 
     @Override
     public Stream<Object> params(Context context) {
-        return Stream.empty();
+        return expression.params(context);
     }
 
     // Projection