+++ src/main/java/io/github/torand/fastersql/statement/SelectStatement.java
@@ -36,7 +36,11 @@ import java.util.stream.Stream;
 import static io.github.torand.fastersql.Command.SELECT;
 import static io.github.torand.fastersql.dialect.Capability.LIMIT_OFFSET;
 import static io.github.torand.fastersql.statement.Helpers.unwrapSuppliers;
-import static io.github.torand.fastersql.util.collection.CollectionHelper.*;
+import static io.github.torand.fastersql.util.collection.CollectionHelper.asList;
+import static io.github.torand.fastersql.util.collection.CollectionHelper.concat;
+import static io.github.torand.fastersql.util.collection.CollectionHelper.isEmpty;
+import static io.github.torand.fastersql.util.collection.CollectionHelper.nonEmpty;
+import static io.github.torand.fastersql.util.collection.CollectionHelper.streamSafely;
 import static io.github.torand.fastersql.util.contract.Requires.require;
 import static io.github.torand.fastersql.util.contract.Requires.requireNonEmpty;
 import static io.github.torand.fastersql.util.functional.Functions.castTo;
@@ -45,7 +49,9 @@ import static io.github.torand.fastersql.util.functional.Predicates.instanceOf;
 import static io.github.torand.fastersql.util.functional.Predicates.not;
 import static io.github.torand.fastersql.util.lang.StringHelper.isBlank;
 import static io.github.torand.fastersql.util.lang.StringHelper.nonBlank;
-import static java.util.Objects.*;
+import static java.util.Objects.isNull;
+import static java.util.Objects.nonNull;
+import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toSet;
 
@@ -237,11 +243,11 @@ public class SelectStatement extends PreparableStatement {
 
         if (nonNull(offset) || nonNull(limit)) {
             if (context.getDialect().supports(LIMIT_OFFSET)) {
-                if (nonNull(limit)) {
-                    sb.append(" limit ?");
-                }
                 if (nonNull(offset)) {
-                    sb.append(" offset ?");
+                    sb.append(" ").append(context.getDialect().formatRowOffsetClause().orElseThrow(() -> new RuntimeException("Dialect " + context.getDialect().getProductName() + " has no row offset clause")));
+                }
+                if (nonNull(limit)) {
+                    sb.append(" ").append(context.getDialect().formatRowLimitClause().orElseThrow(() -> new RuntimeException("Dialect " + context.getDialect().getProductName() + " has no row limit clause")));
                 }
             } else {
                 sb = addLimitOffsetFallback(context, sb, rowFrom(), rowTo());
@@ -268,8 +274,8 @@ public class SelectStatement extends PreparableStatement {
             .orElseThrow(() -> new RuntimeException("Dialect " + context.getDialect().getProductName() + " has no ROWNUM literal"));
 
         if (nonNull(rowFrom) && nonNull(rowTo)) {
-            String limitSql = "select original.*, {ROWNUM} row_no from ( " + innerSql.toString() + " ) original where {ROWNUM} <= ?";
-            String offsetSql = "select * from ( " + limitSql + " ) where row_no >= ?";
+            String limitSql = "select ORIGINAL.*, {ROWNUM} ROW_NO from ( " + innerSql.toString() + " ) ORIGINAL where {ROWNUM} <= ?";
+            String offsetSql = "select * from ( " + limitSql + " ) where ROW_NO >= ?";
             return new StringBuilder(offsetSql.replace("{ROWNUM}", rowNum));
         } else if (nonNull(rowFrom)) {
             String offsetSql = "select * from ( " + innerSql.toString() + " ) where {ROWNUM} >= ?";
@@ -294,17 +300,18 @@ public class SelectStatement extends PreparableStatement {
 
         streamSafely(predicates).flatMap(e -> e.params(context)).forEach(params::add);
 
-        if (nonNull(limit)) {
-            if (context.getDialect().supports(LIMIT_OFFSET)) {
+        if (context.getDialect().supports(LIMIT_OFFSET)) {
+            if (nonNull(offset)) {
+                params.add(offset);
+            }
+            if (nonNull(limit)) {
                 params.add(limit);
-            } else {
+            }
+        } else {
+            if (nonNull(limit)) {
                 params.add(rowTo());
             }
-        }
-        if (nonNull(offset)) {
-            if (context.getDialect().supports(LIMIT_OFFSET)) {
-                params.add(offset);
-            } else {
+            if (nonNull(offset)) {
                 params.add(rowFrom());
             }
         }